Split directory loaders into multiple loaders
We don't know the resource consumption pattern for those loaders yet. And, currently,
the git-checkout loader seem to progress very slowly due to some large origins and
prevents the other loaders to do their work.

I took the opportunity to add limits to the loaders since it got implemented in between.

This also allows to use the dedicated core (tarball), git, mercurial, or svn docker
images instead of the all-in-one save-code-now one.

Refs. swh/infra/sysadm-environment#4931