Install swh-backends and swh applications in the namespace swh
Through the cli.

This is necessary for the log parsing filtering to happen properly (it uses
the namespace to filter data). The namespace used so far, default, was filled
with other pods (whose logs could not be parsed the same way, creating
divergence in what's currently deployed in "production" clusters).

Refs. swh/infra/sysadm-environment#4955