Add new toolbox chart
This is a deployment toolbox pod. It spawns the swh 'toolbox' image with a configuration
to access the remote scheduler of the cluster. For starters, it will allow to register
new scheduler task types.

New tools may be added to it to increase its scope along the way.

Refs. swh/infra/sysadm-environment#5030