swh: Don't start the storage and replayers if a migration is needed
This is a first try for the storage and storage-replayer deployments.
It fails the init of the pod if there is a mismatch between the
db version and the expected code version.

The toolbox could be improved to manually launch the migration, which will
unblock the pods initialisation.

The check can be disabled because it's only available for postgresql backend for now