staging/deposit: Adapt azure configuration to overwrite headers when writing
Currently, the writing in the azure objstorage adds extra headers to the blobs. Which
prevents the correct reading of those blobs when checking the tarballs.

According to the documentation, we can provide some extra headers to happen to avoid
that. From experimenting with the repl on problematic blobs, having no fancy headers
unstuck the reading of the blobs.

Refs. swh/infra/sysadm-environment#5129