swh/ingresses: Fix 502 when services are redeployed
The nginx is not notified as soon a service endpoints are updated so it
sends a couple of requests to a terminated pod.
It's not a problem for idempotent request as nginx retry on another pod
but it's not the case by default for POST requests.

The option is activated by default for all the ingresses because
POST requests are used everywhere

Related to swh/infra/sysadm-environment#5210