production/web: Align archive instance with previous instance moma
Compared to moma, it's currently undersized to 2 replicas, 2 workers, 2 threads so 8
requests handling. As no gunicorn configuration is provided, the default values are set
by the docker image's default values set in the Dockerfile.

In moma, it was configured with 32 workers (with a timeout of 3600s). So we bump the
number of replicas to 4, and the number of workers to 4 as well. That makes up a total
of 32 requests that can be handled, which matches what moma used to do.

This doubles the memory request since we doubled the number of workers, and current use
is near 95% use of memory. This also aligns the request timeout to 3600s.

Refs. swh/infra/sysadm-environment#5110