nixguix/loader: Make the task ack late
As this loader is scheduled like a lister, let's configure it like one. The lister ack
late their messages it ensures once it's scheduled, it's only acknowledged when it's
finally done. That will make those messages resilient to redeployments if any in between
stop the loading.

Refs. swh/infra/sysadm-environment#5213