Declare service monitor to scrape external metrics in clusters
This migrates the equivalent static declarations done in the k8s-cluster-config in this
repository (cassandra for production and staging cluster).

This also adds the scraping of the rabbitmq metrics in the test-staging-rke2. The end
goal being to actually scrape it in staging and production cluster (decommissionning the
ones from pergamon to be able to finally define alertmanager alerts starting with
rabbitmq and then adding other alerts for extra services declared outside the cluster
like kafka, postgresql, ...)

Refs. swh/infra/sysadm-environment#5213