Implement a local pull-through docker cache
The upstream docker registry image only allows caching data from a
single registry. We therefore define a set of upstream registries to
cache data from, and configure a docker-cache StatefulSet for each of
them. For each upstream repository, a httpPrefix is defined to have the
ingress route queries properly.

The pull-through cache can be configured in containerd as a mirror.
containerd degrades gracefully if the mirror is unavailable, pulling
images from the upstream repository instead.

Ref. swh/infra/sysadm-environment#5247