toolbox: Adapt db script for specific backend services
When it's needed (e.g. indexer, scrubber, ...), it allows to specify the module name
(e.g. indexer, scrubber, ...) and the configuration key to introspect the db access
(e.g. indexer_storage, scrubber, ...) in the configuration.

This is especially useful when the configuration key does not match the (top-level)
service type (e.g. indexer). Or when multiple instances of a service exists (e.g. scrubber).

Implementation wise, this adapts the wrapper cli script to allow extra parameters
from the main yaml to be passed along down to the clis.

Refs. swh/infra/sysadm-environment#5253