cluster-components: Invert the alertmanager alert routing logic
As only the alert matching a specific namespace are routed to the irc-relay
today, sending them to the irc relay is acceptable.
We need to remove this constraint to be able to receive global cluster
alerts like the one relative to the cluster status.
If the contraint is remove the irc relay is spammed with the Watchdog
alert. The logic need to be inverted to send all the alerts we only
want to the irc-relay

Related to swh/infra/sysadm-environment#5036
and swh/infra/sysadm-environment#5281