Rework the storage configuration template helpers to be yaml-based
Instead of playing a lot of confusing indentation and whitespace games,
incrementally build configuration data structures using template
functions. Once the structure is built, just output it as yaml, with no
indentation.

When including a template in another, instead of having to pass a
brittle indentation number to the generated template, just parse it with
`fromYaml`, then include it in the data structure.