swh/production: Replay the missing contents in winery
Restart the replay from scratch with check_dst activated.
The errors will be stored in another redis db so we could
compare the differences.
As a new consumer group is used, if it's too long to get
to the current replay position, we could start another instance
of the replayers with the previous consumer group.

Related to swh/infra/sysadm-environment#5295