swh/scrubber: Disable cassandra scrubbers until the statsd configuration can be removed
The statsd configuration is activated globally for an environment
so there is no way to only disable the configuration only for
the cassandra storage scrubbers.

The chart will need a refactoring to allow that.
In the interval, the cassandra scrubbers are fully deactivated in the
meantime

Related to swh/infra/sysadm-environment#5325