kafka & storage replayer: Make the package work
Unfortunately, the current strimzi operator refuses to use partially the authentication.
Once activated, this must be the case for all entities using the kafka topics (be it the
storage or the replayers).

We are not using the authentication part for the storage part so let's disable it for
now.

The storage replayer template got adapted to allow the anonymous mode use.

Refs. swh/infra/sysadm-environment#5344