swh: Add graph deployment template
This declares the template to be able to deploy multiple graph instances in a
kubernetes cluster.

This adds the local-cluster and next-version graph declaration instances.  The
former runs 2 instances, 1 with the sample dataset (stored within the
swh.graph source repository) and another with the python3k dataset.

The latter only runs one instance with the sample dataset so it's fast to
start.

It's possible to make the graph instance retrieves the graph dataset out of
the swh dataset (for test purposes).

Refs. swh/infra/sysadm-environment#5382