components: Reduce false positive ingress alerts
Use only 5XX errors to detect errors ratio.
Using the 4XX status can trigger false positive alerting
for some incomplete object storage or the webapp when
quota are exceeded

As there is now only 500 errors, the threshold can be reduced.
During the last month, it seems a the 10% error rate was
reached when real incident happened.