next-version/graph: Adapt config so ingress comm' is functional
Ingress is usually for cluster external communication to internal cluster
services. We use it in a specific way so we can benefit from the metrics we
already have set for ingress.

But ingress "dns" resolution is not a defacto standard vs the integrated
service resolution so we need to declare externalService so the communication
happens properly.