graph: Make reindexation step happen on persistent volume
The reindexation is still optional. When required, this reindex the data on
the persistent volume. And then the prepares memory volume just as before.

This complicates the orchestration of the init-containers so they happen in
the correct order but that avoids generating some *.ef files which are around
30-40Gb each (graph.ef, graph-transposed.ef).

Refs. swh/infra/sysadm-environment#5448