production/graph-grpc: Decrease requested memory
It's only consuming 10Gi for now of the 500Gi (~2%). It's probably causing
some pressure in the other pods running in the same node too.

If the graph grpc pod is going to use more and the instance is not ok because
of it, the blackbox probe should warn us anyways.

Refs. swh/infra/sysadm-environment#5448