production/graph: Use proper requestedMemory
It's supposedly going up to 500Gi which is too high for rancher-node-metal05 ("only"
502Gi of memory).

By increasing it that way, this should help the kube scheduler not too schedule too much
other pods on the same node while the graph is running. With the current 100Gi, lots of
pods (too many?) are scheduled to run on rancher-node-metal05.

Refs. swh/infra/sysadm-environment#5448