local: Enable swh rpc clients retry behavior in case of failure
For example, when a new deployment of the associated rpc service occurs, the client
connection to the rpc service can fail (with a 502 for example). This allow the client
to retry the connection in a backoff fashion. This should prevent the main proce
depending on the rpc service to crash if this happens.

Refs. swh/infra/sysadm-environment#5493