production: Enable swh rpc clients retry behavior in case of failure
For example, when a new deployment of the associated rpc service occurs, the connection
can fail. This ensures, a retry behavior occurs in the client of such rpc service. This
should prevent the failure of the main (client) process depending on the rpc
service (being redeployed).

Refs. swh/infra/sysadm-environment#5493