loader-deposit: Increase minReplicaCount to reduce downscaling event
Preventing some load-deposit to finish properly.

We limit the case where pods gets warm shutdown because there is almost no
messages left in the queue. As they have 3600s to finish when they receive
that signal, but some deposit ingestion takes longer than this period to
finish [1]

As we activated ackLate to stop losing messages, that means, the message is
still in the queue and will get consumed again by another pod. Each deposit
loading takes apparently the same amount of time for each ingestion of the
same deposit. If the new pod is eventually warmed shutdown again, that could
loop a while before being actually fully ingested (and the message
acknowledged).

[1] some takes up more than 6000s, e.g. this one took 9142.2s
```
{"asctime": "2025-03-20 20:56:35,996", "threadName": "MainThread", "pathname": "/opt/swh/venv/lib/python3.11/site-packages/celery/app/trace.py", "lineno": 128, "funcName": "info", "task_name": null, "task_id": null, "name": "celery.app.trace", "levelname": "INFO", "message": "Task swh.loader.package.deposit.tasks.LoadDeposit[ddb169b3-ac71-4599-87b9-098393817903] succeeded in 9142.2018524101s: {'status': 'eventful', 'snapshot_id': 'ff59853917f5a9b8770c8a2998753c6baabab885'}", "data": {"id": "ddb169b3-ac71-4599-87b9-098393817903", "name": "swh.loader.package.deposit.tasks.LoadDeposit", "return_value": "{'status': 'eventful', 'snapshot_id': 'ff59853917f5a9b8770c8a2998753c6baabab885'}", "runtime": 9142.2018524101, "args": "()", "kwargs": "{'url': 'https://doi.org/10.5281/zenodo.598174', 'deposit_id': 43423}"}}
```

Refs. swh/infra/sysadm-environment#5512