datasets: Keep the debug pod indefinitly
In waiting for future release of swh-graph (blocked by the missing release of its deps
webgraph-rs), we will use the pod in debug mode. As we don't know the time it takes to
do a one-off export, it's better to keep the pod up as long as we need it too to avoid
spurious restarts.

Refs. swh/infra/sysadm-environment#5623