cluster-comp/cnpg: Adapt the folder recovery
That way the recovery should work (famous last words).

```
By default, the recovery method strictly uses the name of the cluster in
the externalClusters section as the name of the main folder of the backup data
within the object store. This name is normally reserved for the name of the
server. You can specify a different folder name using the
barmanObjectStore.serverName property.
```

https://cloudnative-pg.io/documentation/current/appendixes/backup_barmanobjectstore/#recovery-from-an-object-store

Refs. swh/infra/sysadm-environment#5655