cnpg: Allow to determine which postgresql image to use
Currently it was only set for the coar-notify db. This unifies the other dbs to use the
same image. This means the image will be managed by the operator too.

This defaults to use the 17 version.

The next-version dbs were already using one similar version but not the same.

Refs. swh/infra/sysadm-environment#5655