cnpg: Allow to determine which postgresql image to use 2/2
As it was not mentioned before, the default image was used. Let's migrate to use the
image catalog ref instead and fallback to image if not using the catalog ref. That
should ease the argocd sync into migrating properly to image catalog ref.

The operator will be in charge of updating properly the dbs.

Refs. swh/infra/sysadm-environment#5655