swh/staging: adapt second layer of timeout
Without retry, the proxy_read_timeout is now the limit
for long queries.
Let increase it to the same value the the next_upstream_timeout.

It's not optimal because the proxy read timeout should be greater than the
next_upstream_timeout to let the time to retry on another server.
But given the already long timeout, it's not reasonable to double the proxy_read_timeout

Related to swh/infra/sysadm-environment#5684