cluster-comp/pooler: Drop unused pooler code
This (cnpg) pooler implementation is only per postgresql cluster and we want
to be able to have one pooler for all of our dbs.

So we won't be using that part, we will be using one pooler per kubernetes
cluster and open the access to the dbs on that instance. The configuration
will happen through the cluster-configuration/pgbouncer template instead.

Refs. swh/infra/sysadm-environment#5655