cnpg: Properly configure recovery setup per instance
Recovery is disabled by default (enabled to false). The enabled to true allows to enable
it but should not be part of the recovery dict after that.

Previously that added the "enabled: true" in the output yaml while it's not part of the
Cluster specification (resulting in out-of-sync issues).

Refs. swh/infra/sysadm-environment#5655