cluster-configuration/rabbitmq: Migrate to official rabbitmq operator
Unfortunately, the official operator is not packaged through helm and it does not
provide an (git) in-tree yaml file (which provides the crds) [1] They only release new
version through github release and argocd does not provide means to manage this kind of
release.

So we integrate both the cluster-operator-$version.yaml and
messaging-topology-operator-with-certmanager-$version2.yaml in our chart. The argocd
application deployed then clone our swh-charts repository to fetch those files and apply
those per environment.

This commit introduces this new kind of deployment. If this pattern generalizes, it'd be
interesting to move those kind of files in a dedicated repository whose sole goal would
be to centralize those kind of files (with branches management like we currently do).

Another hint would be to look into kustomize as those repositories do use it (that would
mean generate ourselves already released files but... to discuss).

[1] same goes for the other messaging-topology-operator (needed for manaing among other
things exchanges)