rabbitmq: Set container image to use to functional image 4.1.3
Unfortunately, at least 3.10, 3.11, 3.12, ... images are not functional. They are
failing to start the container which fails to write logs on read-only filesystem. No
configuration to disable logs was found to work around this issue, so use the latest
rabbitmq image as bitnami charts used to do (iso functional).

Refs. swh/infra/sysadm-environment#5702